third_party/blink/web_tests/platform/generic/external/wpt/css/css-logical/parsing/border-inline-valid-expected.txt

This is a testharness.js-based test.
PASS e.style['border-inline'] = "1px dotted red" should set the property value
PASS e.style['border-inline'] = "double" should set the property value
PASS e.style['border-inline-start'] = "green double thin" should set the property value
PASS e.style['border-inline-start'] = "green" should set the property value
PASS e.style['border-inline-end'] = "thin" should set the property value
FAIL e.style['border-inline-end'] = "calc(10px - 0.5em) dotted red" should set the property value assert_equals: serialization should be canonical expected "calc(-0.5em + 10px) dotted red" but got "calc(10px - 0.5em) dotted red"
Harness: the test ran to completion.
